% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecnoise.R
\name{ecnoise}
\alias{ecnoise}
\title{Run the ECnoise algorithm on a provided difference stencil}
\usage{
ecnoise(fval)
}
\arguments{
\item{fval}{vector of function evaluations}
}
\value{
list() bundling outputs:
        $fnoise = scalar noise level
        $level  = vector of differences
        $inform = string stating algorithm results
}
\description{
Reference: More' and Wild, "Estimating computational noise"
(2011) SIAM J. Sci. Comput.
}
\details{
Implemented by Zach del Rosario, Feb. 2019
}
\examples{

}
