% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/higham.R
\name{higham}
\alias{higham}
\title{Implementation of the Higham function}
\usage{
higham(t, L = 30)
}
\arguments{
\item{t}{evaluation point
#param L number of iterations; more leads to greater noise}
}
\value{
function value
}
\description{
A form of f(t) = t^2 corrupted by a user-specified level of computational
noise
}
\details{
Reference: More' and Wild, "Estimating computational noise"
(2011) SIAM J. Sci. Comput.

Implemented by Zach del Rosario, Feb. 2019
}
\examples{

}
